# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fashion Xpress API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: Admin - list all bookings, optionally filtered by status
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a home visit booking from the current home visit cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Admin - approve, reject, complete, or cancel a booking
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/reschedule:
    patch:
      operationId: rescheduleBooking
      tags: [bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleInput"
      responses:
        "200":
          description: Rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/assign-executive:
    patch:
      operationId: assignExecutiveToBooking
      tags: [bookings]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignExecutiveInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/products:
    post:
      operationId: addBookingProduct
      tags: [bookings]
      summary: Admin/executive - add a recommended product to a booking
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingProductInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingProduct"
  /bookings/{id}/products/{productId}/status:
    patch:
      operationId: updateBookingProductStatus
      tags: [bookings]
      summary: Executive - mark an assigned product sold, returned, or damaged
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingProductStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingProduct"

  # ---------------- ORDERS ----------------
...
  /bookings/{id}/invoice:
    get:
      operationId: getBookingInvoice
      tags: [invoices]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not generated yet
    post:
      operationId: generateBookingInvoice
      tags: [invoices]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ---------------- ADMIN DASHBOARD ----------------
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Product:
      type: object
      required:
        [
          id,
          name,
          sku,
          categoryId,
          categoryName,
          brandId,
          brandName,
          color,
          sizes,
          fabric,
          occasion,
          mrp,
          sellingPrice,
          stock,
          images,
          rating,
          createdAt,
        ]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        sku: { type: string }
        barcode: { type: ["string", "null"] }
        qrCode: { type: ["string", "null"] }
        categoryId: { type: integer }
        categoryName: { type: string }
        brandId: { type: integer }
        brandName: { type: string }
        color: { type: string }
        sizes: { type: array, items: { type: string } }
        fabric: { type: string }
        occasion: { type: string }
        mrp: { type: number }
        sellingPrice: { type: number }
        purchaseCost: { type: ["number", "null"] }
        supplier: { type: ["string", "null"] }
        stock: { type: integer }
        warehouse: { type: ["string", "null"] }
        rack: { type: ["string", "null"] }
        images: { type: array, items: { type: string } }
        rating: { type: number }
        createdAt: { type: string, format: date-time }
...
    BookingProduct:
      type: object
      required: [id, bookingId, productId, product, status]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        productId: { type: integer }
        product:
          $ref: "#/components/schemas/Product"
        status:
          type: string
          enum: [reserved, sold, returned, damaged]
        priceAtSale: { type: ["number", "null"] }
        isRecommended: { type: boolean }
    BookingProductInput:
      type: object
      required: [productId]
      properties:
        productId: { type: integer }
        isRecommended: { type: boolean, default: true }
    BookingProductStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [reserved, sold, returned, damaged]
        priceAtSale: { type: number }
    Booking:
      type: object
      required:
        [
          id,
          bookingCode,
          customerId,
          name,
          phone,
          email,
          addressText,
          gender,
          preferredDate,
          preferredTime,
          status,
          products,
          createdAt,
        ]
      properties:
        id: { type: integer }
        bookingCode: { type: string }
        customerId: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        addressText: { type: string }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        gender: { type: string }
        age: { type: ["integer", "null"] }
        heightCm: { type: ["number", "null"] }
        weightKg: { type: ["number", "null"] }
        topSize: { type: ["string", "null"] }
        bottomSize: { type: ["string", "null"] }
        preferredFit: { type: ["string", "null"] }
        budget: { type: ["number", "null"] }
        occasion: { type: ["string", "null"] }
        preferredDate: { type: string }
        preferredTime: { type: string }
        preferredColors: { type: array, items: { type: string } }
        preferredBrands: { type: array, items: { type: string } }
        notes: { type: ["string", "null"] }
        status:
          type: string
          enum:
            [
              pending,
              confirmed,
              rescheduled,
              executive_assigned,
              in_progress,
              completed,
              cancelled,
              rejected,
            ]
        executiveId: { type: ["integer", "null"] }
        executiveName: { type: ["string", "null"] }
        products:
          type: array
          items:
            $ref: "#/components/schemas/BookingProduct"
        createdAt: { type: string, format: date-time }
    BookingInput:
      type: object
      required:
        [
          name,
          phone,
          email,
          addressText,
          gender,
          preferredDate,
          preferredTime,
        ]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        addressText: { type: string }
        lat: { type: number }
        lng: { type: number }
        gender: { type: string }
        age: { type: integer }
        heightCm: { type: number }
        weightKg: { type: number }
        topSize: { type: string }
        bottomSize: { type: string }
        preferredFit: { type: string }
        budget: { type: number }
        occasion: { type: string }
        preferredDate: { type: string }
        preferredTime: { type: string }
        preferredColors: { type: array, items: { type: string } }
        preferredBrands: { type: array, items: { type: string } }
        notes: { type: string }
    BookingStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum:
            [
              pending,
              confirmed,
              rescheduled,
              executive_assigned,
              in_progress,
              completed,
              cancelled,
              rejected,
            ]
    RescheduleInput:
      type: object
      required: [preferredDate, preferredTime]
      properties:
        preferredDate: { type: string }
        preferredTime: { type: string }
    AssignExecutiveInput:
      type: object
      required: [executiveId]
      properties:
        executiveId: { type: integer }

    # ---------- EXECUTIVES ----------
...
    Invoice:
      type: object
      required:
        [
          id,
          bookingId,
          invoiceNumber,
          subtotal,
          tax,
          total,
          paymentMethod,
          paymentStatus,
          createdAt,
        ]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        invoiceNumber: { type: string }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        paymentMethod:
          type: string
          enum: [cash, upi, card, wallet]
        paymentStatus:
          type: string
          enum: [pending, paid]
        createdAt: { type: string, format: date-time }
    InvoiceInput:
      type: object
      required: [paymentMethod]
      properties:
        paymentMethod:
          type: string
          enum: [cash, upi, card, wallet]

    # ---------- ADMIN DASHBOARD ----------
...