# lib/api-zod/src/generated/types/rescheduleInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Fashion Xpress API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RescheduleInput {
  preferredDate: string;
  preferredTime: string;
}